Started refactoring process of the whole program
