+++ frontend/src/functions/CourseFunctions.jsx
