+++ frontend/vite.config.js
@@ -4,10 +4,10 @@ import dotenv from 'dotenv';
 
 dotenv.config();
 
+const inStaging = process.env.STAGING === 'true'
 const devPort = process.env.VITE_DEV_PORT ? parseInt(process.env.VITE_DEV_PORT, 10) : 3001;
 console.log('VITE_DEV_PORT:', devPort);
 
-const inStaging = process.env.STAGING === 'true'
 
 // https://vitejs.dev/config/
 export default defineConfig({
