+++ package.json
@@ -2,28 +2,36 @@
   "name": "kurssiesitieto",
   "private": true,
   "version": "0.0.0",
-  "type": "module",
+  "type": "commonjs",
   "scripts": {
-    "dev": "vite",
+    "dev": "NODE_ENV=development concurrently \"nodemon backend/index.js\" \"vite ./frontend\"",
+    "build": "vite build ./frontend --outDir ../backend/dist",
+    "prod": "NODE_ENV=production node backend/index.js",
     "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview",
-    "start": "vite",
-    "test": "vitest",
-    "coverage": "vitest --coverage",
-    "setup-and-run": "rm -r backend/dist || true && cd frontend && npm run build && mv dist ../backend && cd ../backend && npx nodemon index.js",
-    "install:frontend": "cd frontend && npm install",
-    "install:backend": "cd backend && npm install",
-    "install:both": "npm run install:frontend && npm run install:backend",
-    "setup-and-run-windows": "rmdir /s /q backend\\dist && cd frontend && npm run build && move dist ..\\backend && cd ..\\backend && npx nodemon index.js"
+    "coverage": "jest --coverage"
   },
   "dependencies": {
+    "@emotion/react": "^11.11.3",
+    "@emotion/styled": "^11.11.0",
+    "@mui/icons-material": "^5.15.11",
+    "@mui/material": "^5.15.7",
     "axios": "^1.6.5",
+    "concurrently": "^8.2.2",
+    "cors": "^2.8.5",
     "dagre-reactjs": "^1.0.0",
+    "dotenv": "^16.4.1",
+    "express": "^4.18.2",
+    "pg": "^8.11.3",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router-dom": "^6.22.3",
     "reactflow": "^11.10.1"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.23.9",
+    "@babel/preset-react": "^7.23.3",
+    "@testing-library/jest-dom": "^6.4.2",
+    "@testing-library/react": "^14.2.1",
     "@types/react": "^18.2.43",
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.2.1",
@@ -32,6 +40,17 @@
     "eslint-plugin-react": "^7.33.2",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.5",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^29.7.0",
+    "jest-dom": "^4.0.0",
+    "jest-environment-jsdom": "^29.7.0",
+    "nodemon": "^3.0.3",
     "vite": "^5.0.8"
+  },
+  "babel": {
+    "presets": [
+      "@babel/preset-env",
+      "@babel/preset-react"
+    ]
   }
 }
